requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pymongo==4.10.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
